Fix arm32 compilation

+++ source/inochi2d/core/param/binding.d
@@ -766,9 +766,9 @@ public:
     }
 
     T interpolateNearest(vec2u leftKeypoint, vec2 offset) {
-        ulong px = leftKeypoint.x + ((offset.x >= 0.5) ? 1 : 0);
+        size_t px = leftKeypoint.x + ((offset.x >= 0.5) ? 1 : 0);
         if (parameter.isVec2) {
-            ulong py = leftKeypoint.y + ((offset.y >= 0.5) ? 1 : 0);
+            size_t py = leftKeypoint.y + ((offset.y >= 0.5) ? 1 : 0);
             return values[px][py];
         } else {
             return values[px][0];
